Get type-specific column length/precision/scale defaulting working again
And fix resulting breakage to envers

+++ hibernate-core/src/test/java/org/hibernate/test/annotations/embeddables/nested/NestedEmbeddableMetadataTest.java
@@ -51,7 +51,7 @@ public class NestedEmbeddableMetadataTest extends BaseUnitTestCase {
 			Value descriptionValue = investmentMetadata.getProperty( "description" ).getValue();
 			assertEquals( 1, descriptionValue.getColumnSpan() );
 			Column selectable = (Column) descriptionValue.getColumnIterator().next();
-			assertEquals( 500, selectable.getLength() );
+			assertEquals( (Long) 500L, selectable.getLength() );
 			Component amountMetadata = (Component) investmentMetadata.getProperty( "amount" ).getValue();
 			SimpleValue currencyMetadata = (SimpleValue) amountMetadata.getProperty( "currency" ).getValue();
 			CustomType currencyType = (CustomType) currencyMetadata.getType();
